fix nextDate handler, case when err == nil
