+++ internal/handlers/nextDate.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"fmt"
-	"go_final_project/internal/models"
+	"go_final_project/internal/utils"
 	"net/http"
 	"time"
 )
@@ -17,22 +17,35 @@ import (
 // Если входные параметры недействительны или при вычислении возникает ошибка, функция возвращает ответ HTTP 400 Bad Request.
 // В противном случае она устанавливает заголовок "Content-Type" в "text/plain" и выводит результат в формате "%s\n".
 func (h *Handler) NextDate(w http.ResponseWriter, r *http.Request) {
+	h.logger.Infof("next date enter")
+
+	if r == nil {
+		err := fmt.Errorf("request is nil")
+		h.SendErr(w, err, http.StatusBadRequest)
+		return
+	}
+
 	now, err := time.Parse("20060102", r.FormValue("now"))
 	if err != nil {
 		h.SendErr(w, err, http.StatusBadRequest)
 		return
 	}
+
 	date := r.FormValue("date")
 	repeat := r.FormValue("repeat")
+
 	if date == "" || repeat == "" {
+		err = fmt.Errorf("date or repeat is empty")
 		h.SendErr(w, err, http.StatusBadRequest)
 		return
 	}
-	next, err := models.NextDate(now, date, repeat)
+
+	next, err := utils.NextDate(now, date, repeat)
 	if err != nil {
 		h.SendErr(w, err, http.StatusBadRequest)
 		return
 	}
+
 	h.logger.Infof("next date updated")
 	w.Header().Set("Content-Type", "text/plain")
 	fmt.Fprintf(w, "%s\n", next)
