+++ internal/handlers/task.go
@@ -36,10 +36,6 @@ import (
 // По умолчанию, если метод запроса не GET, POST, PUT или DELETE, возвращает ошибку 500 Internal Server Error.
 func (h *Handler) Task(w http.ResponseWriter, r *http.Request) {
 	switch r.Method {
-	// В нашем случае необходимо добавить обработку GET-запроса, который возвратит все параметры задачи по её идентификатору.
-	// Если сейчас нажать на иконку редактирования задачи, появится ошибка.
-	// Исправьте ситуацию — реализуйте обработчик GET-запроса /api/task?id=<идентификатор>.
-	// Запрос должен возвращать JSON-объект со всеми полями задачи.
 	case http.MethodGet:
 		id, err := h.GetID(r)
 		if err != nil {
